refactor(engine): split gateway-actions into per-domain modules
gateway-actions.ts was 1184 lines — essentially one ~1020-line
handleSharedAction switch over ~30 action types. Split into a
gateway-actions/ folder with one module per domain (history, fetch-url,
cron, triggers, goals, scripts, skills, plugins, models), each exporting
a SharedActionHandlers map. index.ts merges them into one registry and
keeps handleSharedAction's signature/behavior identical (returns null for
unknown actions). Pure code movement — no behavior change.
