+++ src/__tests__/fuzz.test.ts
@@ -48,7 +48,7 @@ vi.mock("../storage/cron-store.js", () => ({
 const { classify, TalonError } = await import("../core/errors.js");
 await import("../storage/cron-store.js");
 const { handleSharedAction } =
-  await import("../core/engine/gateway-actions.js");
+  await import("../core/engine/gateway-actions/index.js");
 const { resolveModelName } = await import("../storage/chat-settings.js");
 const { registerClaudeModelsStatic, CLAUDE_MODELS_STATIC } =
   await import("../backend/claude-sdk/models.js");
