+++ src/core/engine/gateway.ts
@@ -19,7 +19,7 @@ import { getActiveCount } from "./dispatcher.js";
 import { getHealthStatus } from "../../util/watchdog.js";
 import { getActiveSessionCount } from "../../storage/sessions.js";
 import { log, logError, logDebug } from "../../util/log.js";
-import { handleSharedAction } from "./gateway-actions.js";
+import { handleSharedAction } from "./gateway-actions/index.js";
 import { handlePluginAction } from "../plugin.js";
 import type { FrontendActionHandler } from "../types.js";
 import { BOT_MESSAGE_ACTIONS, noteBotMessage } from "../soul/taps.js";
